Доработан функционал для работы с таблицей UserEntity

+++ src/main/java/mem/MrPonerYea/PoolControl/repository/UserGroupRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface UserGroupRepository extends JpaRepository<UserGroupEntity, Integer> {
+public interface UserGroupRepository extends JpaRepository<UserGroupEntity, Long> {
 }
